Move ImplicitActiveWorkbookReferenceInspectionTests for QuickFixes into relevant files under RD.QuickFixes
